Remove redundant executable file.
The executable file can be created in the `bin` directory of the
virtualenv with `pip install ./`

This will require a relatively recent version of `pip`. If the command
fails please upgrade `pip` with `pip install -U pip`.

Disclaimer: I have not actually tried this within a conda environment, but
I think that it should work.