poetry: Pin numpy < 1.24
`datashader` up to 1.14.4 was using `numpy.warnings` which got removed in 1.24.
They have fixed this on master, but they have not released a new version yet.
Adding this pin forces us to make a new thalassa release as soon as 1.14.5 gets released,
because if we don't we won't be able to use thalassa with up to date numpy versions.

https://github.com/holoviz/datashader/issues/1158