poetry: Fix licencing metadata
We were using the wrong SPDX identifier in the `licence` field, and as a result
the package metadata was mentioning `Other/Proprietary License` as the trove classifier.
Now the next package version will contain the proper classifier for EUPL v1.2

The GPL3 classifier was just wrong, so we remove it.